sorting extracted
