+++ src/plugins/sorting/sorting.plugin.ts
@@ -12,14 +12,10 @@ import type {
   DimensionRows,
   PluginProviders,
 } from '@type';
+import type { SortingOrder, SortingOrderFunction, ColumnSetEvent } from './sorting.types';
 import { getCellRaw, getColumnByProp } from '../../utils/column.utils';
 import { rowTypes } from '@store';
-
-export type SortingOrder = Record<ColumnProp, Order>;
-type SortingOrderFunction = Record<ColumnProp, CellCompareFunc | undefined>;
-type ColumnSetEvent = {
-  order: SortingOrder;
-};
+import { sortIndexByItems } from './sorting.func';
 
 /**
  * Lifecycle
@@ -239,7 +235,7 @@ export class SortingPlugin extends BasePlugin {
         // row indexes
         const proxyItems = storeService.store.get('proxyItems');
 
-        const newItemsOrder = this.sortIndexByItems(
+        const newItemsOrder = sortIndexByItems(
           [...proxyItems],
           source,
           sortingFunc,
@@ -272,39 +268,6 @@ export class SortingPlugin extends BasePlugin {
     return (prop: ColumnProp, a: DataType, b: DataType): number => {
       return -1 * cmp(prop, a, b);
     };
-  }
-
-  sortIndexByItems(
-    indexes: number[],
-    source: DataType[],
-    sortingFunc: SortingOrderFunction = {},
-  ): number[] {
-    // if no sorting - return unsorted indexes
-    if (Object.entries(sortingFunc).length === 0) {
-      // Unsorted indexes
-      return [...Array(indexes.length).keys()];
-    }
-    //
-    /**
-     * go through all indexes and align in new order
-     * performs a multi-level sorting by applying multiple comparison functions to determine the order of the items based on different properties.
-     */
-    return indexes.sort((a, b) => {
-      for (const [prop, cmp] of Object.entries(sortingFunc)) {
-        const itemA = source[a];
-        const itemB = source[b];
-
-        /**
-         * If the comparison function returns a non-zero value (sorted), it means that the items should be sorted based on the given property. In such a case, the function immediately returns the sorted value, indicating the order in which the items should be arranged.
-         * If none of the comparison functions result in a non-zero value, indicating that the items are equal or should remain in the same order, the function eventually returns 0.
-         */
-        const sorted = cmp?.(prop, itemA, itemB);
-        if (sorted) {
-          return sorted;
-        }
-      }
-      return 0;
-    });
   }
 
   getNextOrder(currentOrder: Order): Order {
