+++ src/plugins/sorting/sorting.types.ts
@@ -1,0 +1,10 @@
+import type { CellCompareFunc, ColumnProp, Order } from '@type';
+
+export type SortingOrder = Record<ColumnProp, Order>;
+export type SortingOrderFunction = Record<
+  ColumnProp,
+  CellCompareFunc | undefined
+>;
+export type ColumnSetEvent = {
+  order: SortingOrder;
+};
